# sql/sem/tree: prevent heap escape of DatumAlloc method inputs