The changes in commit e17d713c made the inputs of `DatumAlloc` methods
always escape to the heap, even in the case where the receiver is not
`nil`. This commit fixes this by explicitly allocating a datum only when
the receive is nil, and copying the input value into it.

Here's one benchmark to show that this fixes the regression in
allocations:

Before:

```
BenchmarkMaterializer/int6/hasNulls=false/useSel=true              1376446 ns/op  357.09 MB/s  1095462 B/op  61462 allocs/op
BenchmarkMaterializer/int6/hasNulls=false/useSel=true/batchSize=1     3876 ns/op  123.84 MB/s     2736 B/op     82 allocs/op
```

After:

```
BenchmarkMaterializer/int6/hasNulls=false/useSel=true              798796 ns/op  615.33 MB/s  603869 B/op  22 allocs/op
BenchmarkMaterializer/int6/hasNulls=false/useSel=true/batchSize=1  3351 ns/op    143.23 MB/s    2256 B/op  22 allocs/op
```

Epic: None

Release note: None