@mgartner 2024-12-16
It looks like it's possible: https://github.com/jordanlewis/gcassert?tab=readme-ov-file#directives

But I don't know how they'd get away with a change to the compiler that causes this to regress. I think only a change to this code would cause the input to escape in the future.





<!-- Sent from Reviewable.io -->